Google Sign in & Log out done. (with navigation)
